% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snr-opt.R
\name{dataGEN}
\alias{dataGEN}
\title{Data Generation}
\usage{
dataGEN(n, p, ecc, SNRdB, model)
}
\arguments{
\item{n}{sample size}

\item{p}{number of variable}

\item{ecc}{error covariance}

\item{SNRdB}{SNR in dB}
}
\value{
A list of y and error
}
\description{
Data Generation
}
