% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snr-opt.R
\name{SNRest}
\alias{SNRest}
\title{SNR estimation}
\usage{
SNRest(ExxT, Ey2)
}
\arguments{
\item{ExxT}{Covariance of x}

\item{Ey2}{Signal power}
}
\value{
A list of N and a estimated
}
\description{
SNR estimation
}
