% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snr-opt.R
\name{SNRopt}
\alias{SNRopt}
\title{SNR optimization}
\usage{
SNRopt(N, a)
}
\arguments{
\item{N}{Noise-to-signal ratio matrix}

\item{a}{scaling factor}
}
\value{
A vector of weights.
}
\description{
SNR optimization
}
