% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snr-opt.R
\name{SNRopt_est}
\alias{SNRopt_est}
\title{SNRopt_est}
\usage{
SNRopt_est(y, x)
}
\arguments{
\item{y}{observations}

\item{x}{parent products}
}
\value{
A list of weight and merged product.
}
\description{
SNRopt_est
}
