% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snr-freq.R
\name{SNRopt_freq}
\alias{SNRopt_freq}
\title{SNRopt_freq}
\usage{
SNRopt_freq(
  y = NULL,
  x,
  mode = "DWT",
  wf = "haar",
  J,
  pad = "zero",
  boundary = "periodic",
  theta = 0.1,
  option = "Est"
)
}
\arguments{
\item{theta}{}
}
\description{
SNRopt_freq
}
