% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snr-wav.R
\name{SNRopt_wav}
\alias{SNRopt_wav}
\title{SNRopt_wav}
\usage{
SNRopt_wav(
  x,
  mode = "DWT",
  wf = "haar",
  J,
  pad = "zero",
  boundary = "periodic",
  theta = 0.1
)
}
\arguments{
\item{theta}{}
}
\description{
SNRopt_wav
}
