% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_gen.R
\name{dataGEN}
\alias{dataGEN}
\title{Data Generation}
\usage{
dataGEN(n, p, ecc, SNRdB, model = "sine")
}
\arguments{
\item{n}{sample size}

\item{p}{number of variable}

\item{ecc}{error covariance}

\item{SNRdB}{SNR in dB}

\item{model}{Sinewave or random noise model}
}
\value{
A list of y, x, and error
}
\description{
Data Generation
}
